# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseStatus": {
				"enum": [
					200,
					403,
					401,
					500,
					404
				],
				"type": "number"
			},
			"ErrorResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/ResponseStatus"
					}
				},
				"required": [
					"error",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Clase": {
				"properties": {
					"dia": {
						"type": "number",
						"format": "double"
					},
					"horaInicio": {
						"type": "number",
						"format": "double"
					},
					"horaFin": {
						"type": "number",
						"format": "double"
					},
					"salon": {
						"type": "string"
					}
				},
				"required": [
					"dia",
					"horaInicio",
					"horaFin",
					"salon"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GrupoState": {
				"enum": [
					0,
					1,
					2,
					3
				],
				"type": "number"
			},
			"Grupo": {
				"properties": {
					"nombre": {
						"type": "string"
					},
					"profesor": {
						"type": "string"
					},
					"maximo": {
						"type": "number",
						"format": "double"
					},
					"disponible": {
						"type": "number",
						"format": "double"
					},
					"clases": {
						"items": {
							"$ref": "#/components/schemas/Clase"
						},
						"type": "array"
					},
					"estado": {
						"$ref": "#/components/schemas/GrupoState"
					}
				},
				"required": [
					"nombre",
					"profesor",
					"maximo",
					"disponible",
					"clases",
					"estado"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Dictionary_Grupo_": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"$ref": "#/components/schemas/Grupo"
				}
			},
			"MateriaState": {
				"enum": [
					0,
					1,
					2
				],
				"type": "number"
			},
			"Materia": {
				"properties": {
					"semestre": {
						"type": "number",
						"format": "double"
					},
					"carrera": {
						"type": "string"
					},
					"codigo": {
						"type": "string"
					},
					"nombre": {
						"type": "string"
					},
					"horas": {
						"type": "number",
						"format": "double"
					},
					"creditos": {
						"type": "number",
						"format": "double"
					},
					"requisitos": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isElectiva": {
						"type": "boolean"
					},
					"grupos": {
						"$ref": "#/components/schemas/Dictionary_Grupo_"
					},
					"estado": {
						"$ref": "#/components/schemas/MateriaState"
					}
				},
				"required": [
					"semestre",
					"codigo",
					"nombre",
					"horas",
					"creditos",
					"requisitos",
					"isElectiva",
					"grupos"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Dictionary_Materia_": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"$ref": "#/components/schemas/Materia"
				}
			},
			"PensumInfo": {
				"properties": {
					"codigo": {
						"type": "string"
					},
					"fechaCaptura": {
						"type": "string",
						"format": "date-time"
					},
					"nombre": {
						"type": "string"
					}
				},
				"required": [
					"codigo",
					"fechaCaptura",
					"nombre"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pensum": {
				"allOf": [
					{
						"properties": {
							"materias": {
								"$ref": "#/components/schemas/Dictionary_Materia_"
							}
						},
						"required": [
							"materias"
						],
						"type": "object"
					},
					{
						"$ref": "#/components/schemas/PensumInfo"
					}
				]
			},
			"UserJWT": {
				"properties": {
					"user": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AccessTokenResponse": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/UserJWT"
					},
					"status": {
						"$ref": "#/components/schemas/ResponseStatus"
					}
				},
				"required": [
					"accessToken",
					"user",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserLoginRequest": {
				"properties": {
					"user": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-typescript",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/materias/addpensum": {
			"post": {
				"operationId": "AddPensum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pensum"
							}
						}
					}
				}
			}
		},
		"/materias/deletepensum/{carrera}": {
			"delete": {
				"operationId": "DeleteAllPensum",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "carrera",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/materias/pensum/{carrera}": {
			"get": {
				"operationId": "GetPensum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Pensum"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "carrera",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/divisist/pensum": {
			"get": {
				"operationId": "GetPensum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ci_session",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "delay",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessTokenResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLoginRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}